Improved typing.

+++ src/main/kotlin/day03/TobogganTrajectory.kt
@@ -5,7 +5,7 @@ class TobogganTrajectory(private val map: Array<CharArray>) {
     private val height = map.size
     private val width = map[0].size
 
-    fun part1(): Long = countTrees(3, 1)
+    fun part1(): Int = countTrees(3, 1)
 
     fun part2(): Long {
         return listOf(Pair(1, 1), Pair(3, 1), Pair(5, 1), Pair(7, 1), Pair(1, 2))
@@ -13,10 +13,10 @@ class TobogganTrajectory(private val map: Array<CharArray>) {
             .fold(1, { acc, count -> acc * count })
     }
 
-    private fun countTrees(right: Int, down: Int): Long {
+    private fun countTrees(right: Int, down: Int): Int {
         var row = 0
         var column = 0
-        var count = 0L
+        var count = 0
         while (row < height) {
             if (map[row][column] == '#') {
                 count++
